refactor: qr code reader page
